Update plantation condition tests
